Flyttat metoder till klasser
Några fungerar inte helt 100%
