+++ Program.cs
@@ -5,65 +5,6 @@ namespace Multifabriken
 {
     class Program
     {
-        //Metod för att välja modell på bilen
-        static int carModel()
-        {
-            Console.WriteLine("Vilken bilmodell vill du ha?");
-            Console.WriteLine("1. Volvo");
-            Console.WriteLine("2. Saab");
-            Console.WriteLine("3. Ford");
-            string modelChoise = Console.ReadLine(); //try catch
-            int model = 0;
-            try{
-                model = Convert.ToInt32(modelChoise);
-            }catch{
-                Console.WriteLine("Ange endast siffervärde på bilen");
-            }
-            return model;
-        }
-
-        //Metod för att välja färg på bilen 
-        static int carColor()
-        {
-            Console.WriteLine("Vilken färg ska bilen ha? Mata in din färgkod i hex värde: ");
-            string colorChoise = Console.ReadLine();
-            int color = 00000000;
-            try {
-                color = Convert.ToInt32(colorChoise);
-            }catch{
-                Console.WriteLine("Skriv endast siffervärde på hexfärgen");
-            }
-            return color;
-        }
-        
-        //Metod för att välja lyx eller interiör
-        static string carInterior()
-        {
-            Console.WriteLine("Välj din inredning:");
-            Console.WriteLine("1. Lyx");
-            Console.WriteLine("2. Standard");
-            string interiorChoise = Console.ReadLine();
-            
-            return interiorChoise;
-        }
-
-        //Metod för att välja godistyp
-        static int candyMenu(){
-            Console.WriteLine("Vilken typ av godis vill du ha?");
-            Console.WriteLine("1. Hallon");
-            Console.WriteLine("2. Lakris");
-            Console.WriteLine("3. Jordgubb");
-            Console.WriteLine("4. Päron");
-            string flavourChoise = Console.ReadLine();//try catch / switch?
-            int flavour = 0;
-            try{
-                flavour = Convert.ToInt32(flavourChoise);
-            }catch{
-                Console.WriteLine("Ange endast siffervärde på menyval");
-            }
-            return flavour;
-        }
-
         //Metod för att hämta mängd på beställning från användaren. 
         static double getAmmount(){
             Console.WriteLine("Hur mycket?");
@@ -75,51 +16,6 @@ namespace Multifabriken
                 Console.WriteLine("Ange mängden i siffror");
             }
             return dAmmount;
-        }
-
-        //Metod för fetthalt på mjölk
-        static int milkFat(){
-            Console.WriteLine("Vilken fetthalt vill du ha? ");
-            Console.WriteLine("(Mellan 2% och 11%)");
-            string fat = Console.ReadLine();
-            int partFat = 0;
-            try{
-                partFat = Convert.ToInt32(fat);
-            }catch{
-                Console.WriteLine("Ange fetthalt i heltal");
-            }
-            if (partFat < 2 || partFat > 11){
-                Console.WriteLine("Fetthalten måste vara mellan 2% och 11%");
-            }
-            return partFat;
-        }
-
-        //metod för längd på rör
-        static double pipeLength() 
-        {
-            Console.WriteLine("Mata in önskad längd (cm): ");
-            double inmatning = Convert.ToDouble(Console.ReadLine());
-            return inmatning;
-        }
-
-        //metod för diameter på rör
-        static double pipeDiameter () 
-        {
-            Console.WriteLine("Mata in önskad diameter (cm): ");
-            double inmatning = Convert.ToDouble(Console.ReadLine());
-            return inmatning;
-        }
-
-        //Metod för material på rör
-        static string pipeMaterial()
-        {
-            //Fixa try catch
-            Console.WriteLine("Vilket material vill du ha?");
-            Console.WriteLine("1. Plast");
-            Console.WriteLine("2. Betong");
-            Console.WriteLine("3. Stål");
-            string inmatning = Console.ReadLine();
-            return inmatning;
         }
         
         static void Main(string[] args)
@@ -143,13 +39,13 @@ namespace Multifabriken
                 {
                     case "1"://Godis
                         double candyInmatning = getAmmount();
-                        candyMenu();
+                        Godis.candyMenu();
                         if (candyInmatning < 1){
                             Console.WriteLine("Minst 1kg godis");
                         }else{
                             double candyWeight = candyInmatning;
                         }
-                        Godis nyttGodis = new Godis(candyMenu(), candyInmatning); //Skapar objekt
+                        Godis nyttGodis = new Godis(Godis.candyMenu(), candyInmatning); //Skapar objekt
                         break;
                     case "2"://Havremjölk
 
@@ -159,7 +55,7 @@ namespace Multifabriken
                         }else{
                             double milkFat = milkInmatning;
                         }
-                        Havremjölk mjölkOrder = new Havremjölk(milkFat(), milkInmatning);
+                        Havremjölk mjölkOrder = new Havremjölk(Havremjölk.milkFat(), milkInmatning);
                         
                         break;
                     case "3"://Rör
@@ -170,8 +66,8 @@ namespace Multifabriken
                     case "4"://Bilar
                         Bilar BilBeställning = new Bilar();
                         BilBeställning.menyModell = carModel();//int värde (meny)
-                        carColor();//int värde (hex)
-                        carInterior();//int värde (meny)
+                        Bilar.carColor();//int värde (hex)
+                        Bilar.carInterior();//int värde (meny)
                         break;
                     case "5"://Skriv ut beställning
                     int i = 1;
